+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.1.0 (2015-06-06)
+
+  * Change return deepElement
+  * Change return element and args
+  * Change php version to >= 5.4
+
 ### 1.0.1 (2015-06-06)
 
   * Add exception in constructor